Use uint32_t sign argument to math error functions
This change is equivalent to the commit
https://github.com/ARM-software/optimized-routines/commit/c65db17340782d647c49e17cbba244862dc38402
and only affects code that is from the Arm optimized-routines project.

It does not affect the observable behaviour, but the code generation
can be different on 64bit targets.  The intention is to make the
portable semantics of the code obvious by using a fixed size type.

+++ newlib/libm/common/math_config.h
@@ -199,10 +199,10 @@ eval_as_double (double x)
 # define unlikely(x) (x)
 #endif
 
-HIDDEN float __math_oflowf (unsigned long);
-HIDDEN float __math_uflowf (unsigned long);
-HIDDEN float __math_may_uflowf (unsigned long);
-HIDDEN float __math_divzerof (unsigned long);
+HIDDEN float __math_oflowf (uint32_t);
+HIDDEN float __math_uflowf (uint32_t);
+HIDDEN float __math_may_uflowf (uint32_t);
+HIDDEN float __math_divzerof (uint32_t);
 HIDDEN float __math_invalidf (float);
 
 /* Shared between expf, exp2f and powf.  */
